Hybride-RAG/requirements.txt
fastapi
markdown
python-dotenv
pypdf
langchain
langchain-core
langchain-community
langchain-huggingface
langchain_text_splitters
pinecone
pinecone-text
uvicorn[standard]
jinja2
python-multipart
pymongo
redis
Pillow
pymupdf
transformers
numpy
aiohttp
cachetools
langsmith
langchain_core
